Substituição type id Cargos por short e correçào de Querys para cadastrar cliente e funcionário

+++ Models/Funcionario.cs
@@ -2,7 +2,7 @@ namespace logiWeb.Models
 {
   public class Funcionario : Pessoa
   {
-    public int IdCargo { get; set; }
+    public short IdCargo { get; set; }
     public Cargo Cargo { get; set; } = new Cargo();
     public string Login { get; set; } = string.Empty;
     public string Senha { get; set; } = string.Empty;
